requirements.txt
fontTools
PyYAML